Create list rentals,people, books modules
